requirements.txt
discord.py
asyncpg
click
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/webkaiyo/parsedatetime
python-dateutil
humanize
jishaku
discord-flags
aioredis
colorama
